requirements.txt
beautifulsoup4==4.12.3
requests==2.31.0
pylint==3.0.3
Flask==3.0.2
python-dotenv==1.0.1
pymongo==4.6.1